+++ CHANGELOG.md
@@ -10,6 +10,8 @@ All notable changes to this project will be documented in this file.
 - Handling of log files with "Claw of the Fallen" as the encounter id.
 - Both bosses are now returned for the "Voice & Claw of the Fallen" strike
   mission.
+- Fixed CM detection for Skorvald logs done after the 2020-09-15 patch
+  (introduction of Sunqua Peak).
 
 ## 0.4.2 - 2020-08-28
 ### Fixed